fix: Bug fixes in model fields with vehicle values

+++ sales/models.py
@@ -15,7 +15,7 @@ class Vend(models.Model):
     vehicle = models.ForeignKey(Vehicle, on_delete=models.CASCADE, verbose_name='Veiculo',)
     date_sale = models.DateTimeField(auto_now_add=True, verbose_name='Data da Venda',)
     payment = models.CharField(choices=PAYMENT_OPTION_STATUS, default='CREDITO', verbose_name='Metodo Pagamento',)
-    value = models.IntegerField(verbose_name='Valor Final',)
+    value = models.DecimalField(max_digits=10, decimal_places=2, verbose_name='Valor Final',)
     update_at = models.DateTimeField(auto_now=True, verbose_name='Atualizado em',)
 
     class Meta:
@@ -33,7 +33,7 @@ class SalesProposal(models.Model):
     status = models.CharField(choices=STATUS_CHOICES, default='AGUARDANDO', verbose_name='Status da Proposta',)
     customer = models.ForeignKey(Customer, on_delete=models.CASCADE, verbose_name='Cliente',)
     vehicle = models.ForeignKey(Vehicle, on_delete=models.CASCADE, verbose_name='Veiculo',)
-    value = models.IntegerField(verbose_name='Valor Final',)
+    value = models.DecimalField(max_digits=10, decimal_places=2, verbose_name='Valor Final',)
     date_proposal = models.DateTimeField(auto_now_add=True, verbose_name='Data da Prosposta',)
     update_at = models.DateTimeField(auto_now=True, verbose_name='Atualizado em',)
